Project_Aether/C-Suite_Active_Logic/CMO/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
google-genai>=1.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1
aiohttp>=3.9.0
